[#25735537] Fix a bug where Tandem's page_helpers weren't available to one of the applications using Tandem on a devise page during the first request to that page. Reloading that page would then work as expected.
I wasn't able to replicate the behavior in our dummy app, but added a request spec anyway mimicking the behavior in the other application. I also verified this fix fixes the other application's bug.

The fix was to use ActiveSupport's lazy loading to add the helpers to both action_controller and action_view. Adding directly to ActionView::Base in addition to ActionController::Base might have also solved the problem. I'm not sure what's preferred, but this seems nice.